requirements.txt
geojson==3.1.0
PyYAML==6.0
Requests==2.32.3
